Removing code replication.
