+++ Makefile.emu
@@ -54,7 +54,7 @@ LDFLAGS += -L. -L$(OUTPUT_PATH)
 # Source file definitions...
 ###########################################################################
 SOURCES_CPP   = src/main-emu.cpp
-SOURCES_CPP  += src/Digitabulum/CPLDDriver/SimCPLDDriver.cpp
+SOURCES_CPP  += src/Digitabulum/CPLDDriver/CPLDDriver.cpp
 SOURCES_CPP  += src/Digitabulum/CPLDDriver/SimSPIBusOp.cpp
 SOURCES_CPP  += src/Digitabulum/LSM9DS1/IIU.cpp
 SOURCES_CPP  += src/Digitabulum/LSM9DS1/LSM9DS1.cpp
