+++ src/Digitabulum/CPLDDriver/CPLDDriver.h
@@ -444,6 +444,7 @@ class CPLDDriver : public EventReceiver, public BusAdapter<SPIBusOp> {
     int8_t attached();      // This is called from the base notify().
     #if defined(MANUVR_CONSOLE_SUPPORT)
       void procDirectDebugInstruction(StringBuilder*);
+      void printHardwareState(StringBuilder*);
     #endif  //MANUVR_CONSOLE_SUPPORT
 
     /* High-level hardware control and discovery. */
@@ -486,6 +487,7 @@ class CPLDDriver : public EventReceiver, public BusAdapter<SPIBusOp> {
     inline uint8_t _irq_offset_byte(int idx) {  return (idx >> 1);     };
     inline uint8_t _irq_offset_bit(int idx) {   return (idx << 2);     };
 
+    void _deinit();
     int8_t readRegister(uint8_t reg_addr);
     int8_t writeRegister(uint8_t reg_addr, uint8_t val);
